+++ crates/bevy_mesh/src/primitives/dim3/sphere.rs
@@ -1,10 +1,7 @@
-use core::f32::consts::PI;
-
-use crate::{
-    mesh::{Indices, Mesh, MeshBuilder, Meshable},
-    render_asset::RenderAssetUsages,
-};
+use crate::{Indices, Mesh, MeshBuilder, Meshable};
+use bevy_asset::RenderAssetUsages;
 use bevy_math::{ops, primitives::Sphere};
+use core::f32::consts::PI;
 use hexasphere::shapes::IcoSphere;
 use thiserror::Error;
 use wgpu::PrimitiveTopology;
