+++ crates/bevy_mesh/src/primitives/dim3/triangle3d.rs
@@ -1,10 +1,7 @@
+use crate::{Indices, Mesh, MeshBuilder, Meshable};
+use bevy_asset::RenderAssetUsages;
 use bevy_math::{primitives::Triangle3d, Vec3};
 use wgpu::PrimitiveTopology;
-
-use crate::{
-    mesh::{Indices, Mesh, MeshBuilder, Meshable},
-    render_asset::RenderAssetUsages,
-};
 
 /// A builder used for creating a [`Mesh`] with a [`Triangle3d`] shape.
 pub struct Triangle3dMeshBuilder {
