+++ crates/bevy_render/Cargo.toml
@@ -71,6 +71,7 @@ bevy_winit = { path = "../bevy_winit", version = "0.15.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.15.0-dev" }
 bevy_tasks = { path = "../bevy_tasks", version = "0.15.0-dev" }
 bevy_image = { path = "../bevy_image", version = "0.15.0-dev" }
+bevy_mesh = { path = "../bevy_mesh", version = "0.15.0-dev" }
 
 # rendering
 image = { version = "0.25.2", default-features = false }
@@ -95,7 +96,6 @@ bytemuck = { version = "1.5", features = ["derive", "must_cast"] }
 downcast-rs = "1.2.0"
 thiserror = "1.0"
 futures-lite = "2.0.1"
-hexasphere = "15.0"
 ddsfile = { version = "0.5.2", optional = true }
 ktx2 = { version = "0.3.0", optional = true }
 # For ktx2 supercompression
